requirements.txt
pytesseract==0.3.1
pillow==7.0.0
pdf2image==1.11.0
